Plugged the annuity calculator to the MortgageAffordabilityService

+++ src/main/java/nl/ing/mortgage/affordability/mortgage/check/MortgageAffordabilityRequest.java
@@ -1,11 +1,9 @@
 package nl.ing.mortgage.affordability.mortgage.check;
 
-import java.math.BigDecimal;
-
 public record MortgageAffordabilityRequest(
-        BigDecimal income,
+        Float income,
         int maturityPeriod,
-        BigDecimal loanValue,
-        BigDecimal homeValue
+        Float loanValue,
+        Float homeValue
 ) {
 }
