+++ src/main/java/nl/ing/mortgage/affordability/mortgage/check/business/rules/HomeValueValidator.java
@@ -2,11 +2,9 @@ package nl.ing.mortgage.affordability.mortgage.check.business.rules;
 
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
-
 @Service
 class HomeValueValidator {
-    public boolean qualifies(BigDecimal homeValue, BigDecimal loanAmount) {
+    public boolean qualifies(Float homeValue, Float loanAmount) {
         return homeValue.compareTo(loanAmount) >= 0;
     }
 }
