+++ src/main/java/nl/ing/mortgage/affordability/mortgage/check/business/rules/MortgageValidator.java
@@ -2,14 +2,12 @@ package nl.ing.mortgage.affordability.mortgage.check.business.rules;
 
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
-
 @Service
 public class MortgageValidator {
     private final HomeValueValidator homeValueValidator = new HomeValueValidator();
     private final IncomeValidator incomeValidator = new IncomeValidator();
 
-    public boolean qualifies(BigDecimal income, BigDecimal homeValue, BigDecimal loanAmount) {
+    public boolean qualifies(Float income, Float homeValue, Float loanAmount) {
         return getHomeValueValidator().qualifies(homeValue, loanAmount) && getIncomeValidator().qualifies(income, loanAmount);
     }
